Split storage and persistence validation
