+++ CMakeLists.txt
@@ -212,6 +212,7 @@ target_sources(
         src/game_data_validation_editor_metadata.c
         src/game_data_validation_grid.c
         src/game_data_validation_input.c
+        src/game_data_validation_persistence.c
         src/game_data_validation_render_components.c
         src/game_data_validation_scene_editor.c
         src/game_data_validation.c
