+++ src/game_data_validation_internal.h
@@ -116,6 +116,8 @@ bool validate_input_assignment_sets(validation_context *ctx, yyjson_val *root);
 bool validate_input_profiles(validation_context *ctx, yyjson_val *root, validation_names *names);
 bool validate_grid_maps(validation_context *ctx, yyjson_val *root);
 bool validate_grid_pickup_layers(validation_context *ctx, yyjson_val *root, validation_names *names);
+bool validate_storage(validation_context *ctx, yyjson_val *root);
+bool validate_persistence(validation_context *ctx, yyjson_val *root, validation_names *names);
 bool is_single_byte_string(yyjson_val *value);
 
 bool is_vec_array(yyjson_val *value, size_t min_count);
